+++ sudachi/src/analysis/mod.rs
@@ -16,13 +16,12 @@
 
 use std::str::FromStr;
 
-use morpheme_list::MorphemeList;
+use morpheme::MorphemeList;
 
 use crate::error::SudachiResult;
 
 pub mod lattice;
 pub mod morpheme;
-pub mod morpheme_list;
 pub mod node;
 pub mod stateless_tokenizer;
